--- src/main/java/cart/auth/AuthorizationExtractor.java
@@ -1,9 +1,0 @@
-package cart.auth;
-
-import javax.servlet.http.HttpServletRequest;
-
-public interface AuthorizationExtractor<T> {
-    String AUTHORIZATION = "Authorization";
-
-    T extract(HttpServletRequest request);
-}
