+++ src/main/java/cart/config/WebMvcConfiguration.java
@@ -1,5 +1,6 @@
 package cart.config;
 
+import cart.auth.AuthenticationPrincipalArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
